feat: group <-> animation

+++ app_animation/views.py
@@ -3,6 +3,7 @@ from django.shortcuts import redirect
 from django.contrib.auth.decorators import login_required
 from .SQL_animation import Animation
 from app_song.SQL_song import Song
+from app_group.SQL_group import Group
 
 
 
@@ -10,13 +11,19 @@ from app_song.SQL_song import Song
 def animations(request):
     error = ''
 
-    selected_group = request.session.get('selected_group')
+    group_selected = ''
+    group_id = request.session.get('group_id', '')
+    url_token = request.session.get('url_token', '')
+    if group_id != '':
+        group = Group.get_group_by_id(group_id, url_token, request.user.username)
+        group_selected = group.name
     
-    if selected_group:
+    if group_selected:
         if request.method == 'POST':
 
             if request.method == 'POST':
                 new_animation = Animation(
+                                group_id = group_id,
                                 name = request.POST.get('txt_new_name'),
                                 description = request.POST.get('txt_new_description'),
                                 date = request.POST.get('dt_new_date'),
@@ -27,7 +34,7 @@ def animations(request):
                 request.POST['txt_new_description'] = ''
                 request.POST['txt_new_date'] = ''
                 
-        animations = Animation.get_all_animations()
+        animations = Animation.get_all_animations(group_id)
     
     else:
         animations = []
@@ -39,6 +46,7 @@ def animations(request):
         'name': request.POST.get('txt_new_name', ''),
         'description': request.POST.get('txt_new_description', ''),
         'date': request.POST.get('txt_new_date', ''),
+        'group_selected': group_selected,
         'error': error,
         })
 
@@ -47,68 +55,84 @@ def animations(request):
 def modify_animation(request, animation_id):
     error = ''
 
-    animation = Animation.get_animation_by_id(animation_id)
-
-    if request.method == 'POST':
-        if 'btn_cancel' not in request.POST:
-            if not animation.name:
-                error = "[ERR2]"
-            else:
-                animation.name = request.POST.get('txt_name')
-                animation.description = request.POST.get('txt_description')
-                animation.date = request.POST.get('dt_date')
-                animation.save()
-
-                if 'btn_new_song' in request.POST:
-                    animation.new_song_verses(request.POST.get('sel_song_id'))
-                
-                for song in animation.songs:
-                    if request.POST.get(f'box_delete_song_{song['animation_song_id']}', 'off') == 'on':
-                        animation.delete_song(song['animation_song_id'])
-                    else:
-                        animation.update_song_num(song['animation_song_id'], request.POST.get(f'lis_move_to_{song['animation_song_id']}'))
-                
-                # verses selected
-                for verse in animation.verses:
-                    animation_song_id = verse['animation_song_id']
-                    verse_id = verse['verse_id']
-                    box_name = f"box_verse_{animation_song_id}_{verse_id}"
-                    
-                    if request.POST.get(box_name, 'off') == 'on':
-                        animation.update_verse_selected(animation_song_id, verse_id, True)
-                    else:
-                        animation.update_verse_selected(animation_song_id, verse_id, False)
-            
-            # reload animation
-            animation = Animation.get_animation_by_id(animation_id)
-
-        if any(key in request.POST for key in ['btn_save_exit', 'btn_cancel']):
-            return redirect('animations')
-    
-        # Recalculate the 'order' for all songs
-        for index, song in enumerate(animation.songs):
-            animation.update_song_num(song['animation_song_id'], (index + 1) * 2)
-        animation.all_songs()
-
-    # import song's lyrics
-    database = ''
+    animation = None
+    songs_already_in = []
     list_lyrics = []
-    for song in animation.songs:
-        song_lyrics = Song.get_song_by_id(song['song_id'])
-        song_lyrics.get_verses()
+    group_selected = ''
+    group_id = request.session.get('group_id', '')
+    url_token = request.session.get('url_token', '')
+    if group_id != '':
+        group = Group.get_group_by_id(group_id, url_token, request.user.username)
+        group_selected = group.name
+    
+    if group_selected:
+        animation = Animation.get_animation_by_id(animation_id, group_id)
+        if not animation:
+            return redirect('animations')
+
+        if request.method == 'POST':
+            if 'btn_cancel' not in request.POST:
+                if not animation.name:
+                    error = "[ERR2]"
+                else:
+                    animation.name = request.POST.get('txt_name')
+                    animation.description = request.POST.get('txt_description')
+                    animation.date = request.POST.get('dt_date')
+                    animation.save()
+
+                    if 'btn_new_song' in request.POST:
+                        animation.new_song_verses(request.POST.get('sel_song_id'))
+                    
+                    for song in animation.songs:
+                        if request.POST.get(f'box_delete_song_{song['animation_song_id']}', 'off') == 'on':
+                            animation.delete_song(song['animation_song_id'])
+                        else:
+                            animation.update_song_num(song['animation_song_id'], request.POST.get(f'lis_move_to_{song['animation_song_id']}'))
+                    
+                    # verses selected
+                    for verse in animation.verses:
+                        animation_song_id = verse['animation_song_id']
+                        verse_id = verse['verse_id']
+                        box_name = f"box_verse_{animation_song_id}_{verse_id}"
+                        
+                        if request.POST.get(box_name, 'off') == 'on':
+                            animation.update_verse_selected(animation_song_id, verse_id, True)
+                        else:
+                            animation.update_verse_selected(animation_song_id, verse_id, False)
+                
+                # reload animation
+                animation = Animation.get_animation_by_id(animation_id, group_id)
+
+            if any(key in request.POST for key in ['btn_save_exit', 'btn_cancel']):
+                return redirect('animations')
         
-        full_title = song_lyrics.full_title
-        list_lyrics.append({
-            'song_id': song['song_id'],
-            'full_title': song_lyrics.full_title,
-            'lyrics':  song_lyrics.get_lyrics(),
-        })
+            # Recalculate the 'order' for all songs
+            for index, song in enumerate(animation.songs):
+                animation.update_song_num(song['animation_song_id'], (index + 1) * 2)
+            animation.all_songs()
+
+        # import song's lyrics
+        database = ''
+        list_lyrics = []
+        for song in animation.songs:
+            song_lyrics = Song.get_song_by_id(song['song_id'])
+            song_lyrics.get_verses()
+            
+            full_title = song_lyrics.full_title
+            list_lyrics.append({
+                'song_id': song['song_id'],
+                'full_title': song_lyrics.full_title,
+                'lyrics':  song_lyrics.get_lyrics(),
+            })
+        
+        songs_already_in = animation.get_songs_already_in()
 
     return render(request, 'app_animation/modify_animation.html', {
         'animation': animation,
         'all_songs': Song.get_all_songs(),
-        'songs_already_in': animation.get_songs_already_in(),
+        'songs_already_in': songs_already_in,
         'list_lyrics': list_lyrics,
+        'group_selected': group_selected,
         'error': error,
     })
 
@@ -117,14 +141,26 @@ def modify_animation(request, animation_id):
 def delete_animation(request, animation_id):
     error = ''
 
-    animation = Animation.get_animation_by_id(animation_id)
+    animation = None
+    group_selected = ''
+    group_id = request.session.get('group_id', '')
+    url_token = request.session.get('url_token', '')
+    if group_id != '':
+        group = Group.get_group_by_id(group_id, url_token, request.user.username)
+        group_selected = group.name
+    
+    if group_selected:
+        animation = Animation.get_animation_by_id(animation_id, group_id)
+        if not animation:
+            return redirect('animations')
 
-    if request.method == 'POST':
-        if 'btn_delete' in request.POST:
-            animation.delete()
-        return redirect('animations')
+        if request.method == 'POST':
+            if 'btn_delete' in request.POST:
+                animation.delete()
+            return redirect('animations')
 
     return render(request, 'app_animation/delete_animation.html', {
         'animation': animation,
+        'group_selected': group_selected,
         'error': error,
     })
